:recycle: chore(api) Challenge instead of Tube in CertifConsolidatedFramework
