+++ api/src/certification/configuration/infrastructure/repositories/consolidated-framework-repository.js
@@ -1,7 +1,5 @@
-import { LOCALE } from '../../../../shared/domain/constants.js';
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
-import * as challengeRepository from '../../../../shared/infrastructure/repositories/challenge-repository.js';
-import * as skillRepository from '../../../../shared/infrastructure/repositories/skill-repository.js';
+import { CertificationFrameworksChallenge } from '../../domain/models/CertificationFrameworksChallenge.js';
 import { ConsolidatedFramework } from '../../domain/models/ConsolidatedFramework.js';
 
 export async function create({ complementaryCertificationKey, challenges, version }) {
@@ -19,30 +17,82 @@ export async function create({ complementaryCertificationKey, challenges, versio
 export async function getCurrentFrameworkByComplementaryCertificationKey({ complementaryCertificationKey }) {
   const knexConn = DomainTransaction.getConnection();
 
-  const currentFrameworkChallenges = await knexConn('certification-frameworks-challenges')
+  const currentFrameworkChallengesDTO = await knexConn('certification-frameworks-challenges')
+    .select(
+      'alpha as discriminant',
+      'delta as difficulty',
+      'challengeId',
+      'createdAt',
+      'calibrationId',
+      'complementaryCertificationKey',
+    )
     .where({
       createdAt: knexConn('certification-frameworks-challenges')
         .select('createdAt')
         .orderBy('createdAt', 'desc')
         .first(),
+      complementaryCertificationKey,
     })
     .select('*');
 
-  const currentChallengeIds = currentFrameworkChallenges.map((challenge) => challenge.challengeId);
-  const currentChallenges = await challengeRepository.getMany(currentChallengeIds, LOCALE.FRENCH_SPOKEN);
-
-  const currentSkillIds = currentChallenges.map((challenge) => challenge.skill.id);
-  const currentSkills = await skillRepository.findByRecordIds(currentSkillIds);
-
-  const currentTubeIds = currentSkills.map((skill) => skill.tubeId);
-
-  return _toDomain({
-    complementaryCertificationKey,
-    createdAt: currentFrameworkChallenges[0].createdAt,
-    tubeIds: currentTubeIds,
-  });
+  return _toDomain({ certificationFrameworksChallengesDTO: currentFrameworkChallengesDTO });
 }
 
-function _toDomain({ complementaryCertificationKey, createdAt, tubeIds }) {
-  return new ConsolidatedFramework({ complementaryCertificationKey, createdAt, tubeIds });
+/**
+ * @param {Object} params
+ * @param {Date} params.createdAt
+ * @param {ComplementaryCertificationKeys} params.complementaryCertificationKey
+ * @returns {Promise<ConsolidatedFramework>}
+ */
+export async function findByCreationDateAndComplementaryKey({ createdAt, complementaryCertificationKey }) {
+  const knexConn = DomainTransaction.getConnection();
+
+  const certificationFrameworksChallengesDTO = await knexConn('certification-frameworks-challenges')
+    .select('alpha as discriminant', 'delta as difficulty', 'challengeId', 'createdAt', 'complementaryCertificationKey')
+    .where({
+      complementaryCertificationKey,
+      createdAt,
+    })
+    .orderBy('challengeId');
+
+  if (certificationFrameworksChallengesDTO.length == 0) {
+    return [];
+  }
+
+  return _toDomain({ certificationFrameworksChallengesDTO });
+}
+
+/**
+ * @param {Array<ConsolidatedFramework>} consolidatedFramework
+ * @returns {Promise<void>}
+ */
+export async function save(consolidatedFramework) {
+  const knexConn = DomainTransaction.getConnection();
+
+  for (const calibratedChallenge of consolidatedFramework.challenges) {
+    await knexConn('certification-frameworks-challenges')
+      .update({
+        alpha: calibratedChallenge.discriminant,
+        delta: calibratedChallenge.difficulty,
+        calibrationId: consolidatedFramework.calibrationId,
+      })
+      .where({
+        complementaryCertificationKey: consolidatedFramework.complementaryCertificationKey,
+        createdAt: consolidatedFramework.createdAt,
+        challengeId: calibratedChallenge.challengeId,
+      });
+  }
+}
+
+function _toDomain({ certificationFrameworksChallengesDTO }) {
+  const { complementaryCertificationKey, createdAt, calibrationId } = certificationFrameworksChallengesDTO[0];
+
+  return new ConsolidatedFramework({
+    complementaryCertificationKey,
+    createdAt,
+    calibrationId,
+    challenges: certificationFrameworksChallengesDTO.map(
+      (challenge) => new CertificationFrameworksChallenge(challenge),
+    ),
+  });
 }
